adding bindings
